+++ src/main/java/kang/tableorder/repository/TablesRepository.java
@@ -10,21 +10,21 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TablesRepository extends JpaRepository<TablesEntity, Integer> {
+public interface TablesRepository extends JpaRepository<TablesEntity, Long> {
 
   List<TablesEntity> findAllByRestaurantEntity(RestaurantEntity restaurantEntity);
 
-  Optional<TablesEntity> findByIdAndRestaurantEntity(Integer id, RestaurantEntity restaurantEntity);
+  Optional<TablesEntity> findByIdAndRestaurantEntity(Long id, RestaurantEntity restaurantEntity);
 
   Optional<TablesEntity> findByTabletMacId(String tabletMacId);
 
-  boolean existsByNumberAndRestaurantEntity(Integer number, RestaurantEntity restaurantEntity);
+  boolean existsByNumberAndRestaurantEntity(int number, RestaurantEntity restaurantEntity);
 
   boolean existsByTabletMacId(String tabletMacId);
 
   @Modifying
   @Query("DELETE FROM TABLES t WHERE t.id = :id AND t.restaurantEntity = :restaurantEntity")
-  void deleteByIdAndRestaurantEntity(Integer id, RestaurantEntity restaurantEntity);
+  void deleteByIdAndRestaurantEntity(Long id, RestaurantEntity restaurantEntity);
 
   @Modifying
   @Query("DELETE FROM TABLES t WHERE t.restaurantEntity = :restaurantEntity")
